+++ composer.json
@@ -18,7 +18,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "webignition\\Tests\\CssValidatorOutput\\": "tests/"
+            "webignition\\CssValidatorOutput\\Parser\\Tests\\": "tests/"
         }
     },
     "scripts": {
